requirements.txt
motor
python-telegram-bot==20.8
python-dotenv
